refactor EnvConfig to new file
